+++ cypress.config.js
@@ -1,0 +1,17 @@
+const { defineConfig } = require('cypress')
+
+module.exports = defineConfig({
+  video: false,
+  defaultCommandTimeout: 8000,
+  viewportWidth: 1280,
+  viewportHeight: 1000,
+  retries: 1,
+  e2e: {
+    // We've imported your old cypress plugins here.
+    // You may want to clean this up later by importing these.
+    setupNodeEvents(on, config) {
+      return require('./cypress/plugins/index.js')(on, config)
+    },
+    baseUrl: 'http://localhost:8080',
+  },
+})
