--- cypress.json
@@ -1,8 +1,0 @@
-{
-  "baseUrl": "http://localhost:8080",
-  "video": false,
-  "defaultCommandTimeout": 8000,
-  "viewportWidth": 1280,
-  "viewportHeight": 1000,
-  "retries": 1
-}
