+++ cypress/e2e/gpu-dependant/cross-section-view.cy.js
