+++ package.json
@@ -61,7 +61,7 @@
     "babel-loader": "^8.0.5",
     "copy-webpack-plugin": "^11.0.0",
     "css-loader": "^6.8.1",
-    "cypress": "^5.6.0",
+    "cypress": "^13.6.3",
     "cypress-image-snapshot": "^3.1.1",
     "enzyme": "^3.9.0",
     "enzyme-adapter-react-16": "^1.10.0",
